requirements.txt
transformers>=4.30.0
torch>=2.0.0
pandas>=2.0.0
numpy>=1.24.0
newsapi-python>=0.2.6
tqdm>=4.65.0
scikit-learn>=1.3.0
seaborn>=0.12.0
matplotlib>=3.7.0